// src/app/models/dept-staff.model.ts
export interface DeptStaff {
  userId: number;
  salutationName: string;
  userFName: string;
  userMname: string;
  userLName: string;
  userMobileNumber: string;
  userEmailOfficial: string;
  userProfilepic: string;
  departmentName: string;
  departmentId: number;
  designationName: string;
  useremployeecode: string;
  userGendername?: string;
  userdateofjoin?: string;
  staffTypeName?: string;
  usercategoryName?: string;
  emptype?: string;
  usershortcode?: string;
}